Isolate old governance config
